--- students/k33422/Kirill_glushkov/lab1/alembic/alembic.ini
@@ -1,3 +1,0 @@
-[alembic]
-sqlalchemy.url = postgresql://postgres:postgres@localhost/finance_db
-script_location = alembic
