+++ students/k33422/Kirill_glushkov/lab1/app/database.py
@@ -1,73 +1,19 @@
-from sqlalchemy import create_engine, Column, DateTime, Float, Integer, String, ForeignKey
+from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import sessionmaker, relationship
+from sqlalchemy.orm import sessionmaker
 
 import os
 from dotenv import load_dotenv
-
+Base = declarative_base()
 load_dotenv()
 db_url = os.getenv('SQLALCHEMY_DATABASE_URL')
 
 engine = create_engine(db_url, echo=True)
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
-Base = declarative_base()
-
-class UserDB(Base):
-    __tablename__ = "users"
-
-    id = Column(Integer, primary_key=True, index=True)
-    username = Column(String, unique=True, index=True)
-    email = Column(String, unique=True, index=True)
-    password = Column(String)
-    incomes = relationship("IncomeDB", back_populates="user")
-    expenses = relationship("ExpenseDB", back_populates="user")
-    budgets = relationship("BudgetDB", back_populates="user")
-
-class IncomeDB(Base):
-    __tablename__ = 'incomes'
-
-    id = Column(Integer, primary_key=True)
-    amount = Column(Integer)
-    category_id = Column(Integer, ForeignKey("categories.id"))
-    category = relationship("CategoryDB", back_populates="incomes")
-    user_id = Column(Integer, ForeignKey('users.id'))
-    user = relationship("UserDB", back_populates="incomes")
-
-class ExpenseDB(Base):
-    __tablename__ = 'expenses'
-
-    id = Column(Integer, primary_key=True)
-    amount = Column(Integer)
-    category_id = Column(Integer, ForeignKey("categories.id"))
-    user_id = Column(Integer, ForeignKey('users.id'))
-    user = relationship("UserDB", back_populates="expenses")
-    category = relationship("CategoryDB", back_populates="expenses")
-
-
-class BudgetDB(Base):
-    __tablename__ = 'budgets'
-
-    id = Column(Integer, primary_key=True)
-    category_id = Column(Integer, ForeignKey("categories.id"))
-    amount = Column(Integer)
-    user_id = Column(Integer, ForeignKey('users.id'))
-    user = relationship("UserDB", back_populates="budgets")
-    category = relationship("CategoryDB", back_populates="budgets")
-
-class CategoryDB(Base):
-    __tablename__ = "categories"
-
-    id = Column(Integer, primary_key=True, index=True)
-    name = Column(String, index=True)
-    # Определение связей многие-ко-многим (many-to-many) с другими таблицами
-    incomes = relationship("IncomeDB", back_populates="category")
-    expenses = relationship("ExpenseDB", back_populates="category")
-    budgets = relationship("BudgetDB", back_populates="category")
-
-def init_db():
-    Base.metadata.create_all(bind=engine)
-
 def get_session():
     with SessionLocal() as session:
         yield session
+
+def init_db():
+    Base.metadata.create_all(bind=engine)
